# Better error handling