**Description**: 
This PR exports some errors as global variables to easier check for errors when calling Get* methods.

For now I exported just two errors, key path not found and unknown value type. If you think this type of error management is good I can change it also for other errors where it would make sense.

**Benchmark before change**:
`enchmarkJsonParserLarge-8 	  100000	     73036 ns/op	      32 B/op	       2 allocs/op`
`BenchmarkJsonParserMedium-8	  500000	     12253 ns/op	      32 B/op	       2 allocs/op`
`BenchmarkJsonParserSmall-8 	 5000000	      1217 ns/op	       0 B/op	       0 allocs/op`

**Benchmark after change**:
`BenchmarkJsonParserLarge-8 	  100000	     74244 ns/op	       0 B/op	       0 allocs/op`
`BenchmarkJsonParserMedium-8	  500000	     12597 ns/op	       0 B/op	       0 allocs/op`
`BenchmarkJsonParserSmall-8 	 5000000	      1145 ns/op	       0 B/op	       0 allocs/op`