@ikruglov 2016-04-04
oh, nice. I was thinking about exactly the same patch 👍