@daboyuka 2016-04-04
I was thinking about this as well, thanks for writing it up!

(Though I still think "key path not found" shouldn't be an error :smile:).