@ikruglov 2016-04-04
@daboyuka ahm, i using jsonparser only to check presence of a particular key in a document. Without "key path not found" being a global variable I have only one option to check that the key is missing - to inspect error's string. Which is not nice :-)