@daboyuka 2016-04-04
@ikruglov of course, your PR is perfect as it is. My comment was in reference to my earlier comments in #26. I still feel it's cleaner to handle key-not-found via the `dataType` return rather than as an error, since it's not really an exceptional circumstance, but I'm less sure of this now. In any case, though, this PR isn't the place for tackling the issue.