@buger 2016-04-05
Thank you! I converted rest of the errors to constants as well.