@buger 2016-04-05
This allocations also helped locate bug when `ArrayEach` was throwing error when all data was already processed.